float

+++ rangers_shop/models.py
@@ -145,7 +145,7 @@ class ProdOrder(db.Model):
     def set_price(self, price, quantity):
 
         quantity = int(quantity)
-        price = int(price)
+        price = float(price)
 
         self.price = quantity * price
         return self.price
